dom/worklet/tests/file_basic.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Worklet</title>
  <script type="application/javascript" src="common.js"></script>
</head>
<body>

<script type="application/javascript">

setupTest();

var worklet = window.createWorklet();
ok(!!worklet, "We have a Worklet");

// First loading
worklet.import("common.js")
.then(() => {
  ok(true, "Import should load a resource.");
})

// Second loading - same file
.then(() => {
  return worklet.import("common.js")
})
.then(() => {
  ok(true, "Import should load a resource.");
})

// 3rd loading - a network error
.then(() => {
  return worklet.import("404.js");
})
.then(() => {
  ok(false, "The loading should fail.");
}, () => {
  ok(true, "The loading should fail.");
})

// 4th loading - a network error
.then(() => {
  return worklet.import("404.js");
})
.then(() => {
  ok(false, "The loading should fail.");
}, () => {
  ok(true, "The loading should fail.");
})

// done
.then(() => {
  SimpleTest.finish();
});

</script>
</body>
</html>
